+++ client/src/App.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import './App.css';
-import Search from './components/search-weather/weather';
-import SideNav from './components/side-nav/side-nav';
-import Library from './components/my-library/my-library';
+import Search from './components/Weather/Weather';
+import Nav from './components/Navigation/Nav';
+import Library from './components/MyLibrary/MyLibrary';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 function App() {
   return (
     <BrowserRouter>
-      <SideNav/>
+      <Nav/>
     <Routes>
       <Route path='/search' element={<Search/>} />
       <Route path='/library' element={<Library/>} />
