+++ client/src/components/MyLibrary/MyLibrary.js
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react';
-import './my-library.css';
+import './MyLibrary.css';
 
 export default function MyLibrary() {
