+++ client/src/components/Weather/Weather.js
@@ -1,14 +1,14 @@
 import React, {useEffect, useState} from 'react';
-import './weather.css';
-import Forecast from './forecast';
+import './Weather.css';
+import Forecast from './Forecast';
 import axios from 'axios';
-import Snow from '../assets/images/snow.png';
-import Clouds from '../assets/images/cloudy.png';
-import Rain from '../assets/images/raining.png';
-import Sun from '../assets/images/sunny.png';
-import Clear from '../assets/images/clear.png';
+import Snow from '../Assets/Images/snow.png';
+import Clouds from '../Assets/Images/cloudy.png';
+import Rain from '../Assets/Images/raining.png';
+import Sun from '../Assets/Images/sunny.png';
+import Clear from '../Assets/Images/clear.png';
 
-export default function CurrentWeather() {
+export default function Weather() {
   const [forecastData, setForecastData] = useState();
   const appid = process.env.REACT_APP_API_KEY;
   const [location, setLocation] = useState('');
